Update OWM_EPD47_epaper_v2.5.ino
When value max_readings is changed from the default value 24 to 32 or 40, the days on the x-axes of the graphs at the bottom of the screen, aren't changed. Therefore i added a integer variable (forecast_days) wich is 1 to 5 related to max_readings. In void DrawGraph, line 1027, 1028 and 1029 is changed to use the new forecast_days value and display the right amount of days on the x-axis of the graphs.